talk updates
