+++ src/index.html
@@ -87,7 +87,7 @@
 	<h1>ES6+ and TypeScript: What You Need to Know</h1>
 	<h2 id="greeting">[ TODO: Replace Me ]</h2>
 	<button type="button" name="click_me" id="click_me" disabled>Click Me!</button>
-	<button type="button" name="load_articles" id="load_articles" disabled>Load Articles</button>
+	<!-- <button type="button" name="load_articles" id="load_articles" disabled>Load Articles</button> -->
 
 	<div id="articles"></div>
