+++ src/index.ts
@@ -1,5 +1,4 @@
-import * as moment from 'moment';
-import { Article, NewsSource, requestNews } from './request';
+// import { NewsSource, renderArticle, requestNews } from './article';
 import { delay } from './util';
 
 /**
@@ -25,39 +24,16 @@ function clickMeClick(evt: MouseEvent) {
 		});
 }
 
-function loadArticlesClick(evt: MouseEvent) {
-	requestNews(NewsSource.VERGE, 'top', (articles) => {
-		const articlesDiv = document.querySelector('#articles');
-		articles.forEach((article) => {
-			articlesDiv!.appendChild(renderArticle(article));
-		});
-	}, (error) => {
-		console.error('[ERROR]', error);
-	});
-}
-
-function renderArticle(article: Article): HTMLElement {
-	const div = document.createElement('div');
-	div.classList.add('article');
-	div.innerHTML = `
-	<div class="image">
-		<img src="${article.urlToImage}" />
-	</div>
-	<div class="content">
-		<div class='header'>
-			<a href="${article.url}">${article.title}</a>
-		</div>
-		<div class="meta date">
-			${moment(new Date(article.publishedAt)).startOf('day').fromNow()}
-		</div>
-		<div class="meta description">
-			<p>${article.description}</p>
-		</div>
-	</div>
-	`;
-
-	return div;
-}
+// function loadArticlesClick(evt: MouseEvent) {
+// 	requestNews(NewsSource.ARS, 'top', (articles) => {
+// 		const articlesDiv = document.querySelector('#articles');
+// 		articles.forEach((article) => {
+// 			articlesDiv!.appendChild(renderArticle(article));
+// 		});
+// 	}, (error) => {
+// 		console.error('[ERROR]', error);
+// 	});
+// }
 
 /**
  * Function that initialises the application
@@ -67,16 +43,7 @@ export function init() {
 	clickMeButton.addEventListener('click', clickMeClick);
 	clickMeButton.removeAttribute('disabled');
 
-	const loadArticlesButton = document.querySelector('#load_articles') as HTMLButtonElement;
-	loadArticlesButton.addEventListener('click', loadArticlesClick);
-	loadArticlesButton.removeAttribute('disabled');
-
-	// spawn(function *() {
-	// 	const response = yield fetch(' https://newsapi.org/v1/articles?source=hacker-news&sortBy=top&apiKey=2b88a66af2114afcb1f736a302d51998');
-	// 	const { articles } = yield response.json();
-	// 	articles.forEach((article: any) => {
-	// 		const { title } = article;
-	// 		console.log('article:', title);
-	// 	});
-	// });
+	// const articlesButton = document.getElementById('load_articles') as HTMLButtonElement;
+	// articlesButton.addEventListener('click', loadArticlesClick);
+	// articlesButton.removeAttribute('disabled');
 }
